Pull audit pivot to concern
